Refactor around function graph
New fns convert_space_sliced and convert_space_chunked for operating
efficiently on large pixel buffers without many function lookups. This
is also necessary for more sane implementations of further colorspaces.

Unfortunately this has the side effect that the per-pixel
`convert_space` fn is much slower as it needs to compute the graph for
each pixel. AFAIK the amount of situations where speed is paramount and
space isn't converted across the entire buffer at once is mostly just my
own Pixelbuster which is probably adjustable to this change.